# Conflicts:
#	zeppelin-web/src/app/notebook/paragraph/paragraph.controller.js

+++ spark/src/main/java/org/apache/zeppelin/spark/PySparkInterpreter.java
@@ -42,6 +42,7 @@ import org.apache.commons.exec.ExecuteResultHandler;
 import org.apache.commons.exec.ExecuteWatchdog;
 import org.apache.commons.exec.PumpStreamHandler;
 import org.apache.commons.exec.environment.EnvironmentUtils;
+import org.apache.commons.lang.StringUtils;
 import org.apache.spark.SparkConf;
 import org.apache.spark.api.java.JavaSparkContext;
 import org.apache.spark.sql.SQLContext;
@@ -62,7 +63,7 @@ import py4j.GatewayServer;
  *
  */
 public class PySparkInterpreter extends Interpreter implements ExecuteResultHandler {
-  Logger logger = LoggerFactory.getLogger(PySparkInterpreter.class);
+  private static final Logger LOGGER = LoggerFactory.getLogger(PySparkInterpreter.class);
   private GatewayServer gatewayServer;
   private DefaultExecutor executor;
   private int port;
@@ -74,6 +75,8 @@ public class PySparkInterpreter extends Interpreter implements ExecuteResultHand
   boolean pythonscriptRunning = false;
   private static final int MAX_TIMEOUT_SEC = 10;
   private long pythonPid;
+
+  private IPySparkInterpreter iPySparkInterpreter;
 
   public PySparkInterpreter(Properties property) {
     super(property);
@@ -105,15 +108,46 @@ public class PySparkInterpreter extends Interpreter implements ExecuteResultHand
       throw new InterpreterException(e);
     }
 
-    logger.info("File {} created", scriptPath);
+    LOGGER.info("File {} created", scriptPath);
   }
 
   @Override
   public void open() {
+    // try IPySparkInterpreter first
+    iPySparkInterpreter = getIPySparkInterpreter();
+    if (property.getProperty("zeppelin.pyspark.useIPython", "true").equals("true") &&
+        iPySparkInterpreter.checkIPythonPrerequisite()) {
+      try {
+        iPySparkInterpreter.open();
+        if (InterpreterContext.get() != null) {
+          // don't print it when it is in testing, just for easy output check in test.
+          InterpreterContext.get().out.write(("IPython is available, " +
+              "use IPython for PySparkInterpreter\n")
+              .getBytes());
+        }
+        LOGGER.info("Use IPySparkInterpreter to replace PySparkInterpreter");
+        return;
+      } catch (Exception e) {
+        LOGGER.warn("Fail to open IPySparkInterpreter", e);
+      }
+    }
+    iPySparkInterpreter = null;
+
+    if (property.getProperty("zeppelin.pyspark.useIPython", "true").equals("true")) {
+      // don't print it when it is in testing, just for easy output check in test.
+      try {
+        InterpreterContext.get().out.write(("IPython is not available, " +
+            "use the native PySparkInterpreter\n")
+            .getBytes());
+      } catch (IOException e) {
+        LOGGER.warn("Fail to write InterpreterOutput", e);
+      }
+    }
+
     // Add matplotlib display hook
     InterpreterGroup intpGroup = getInterpreterGroup();
     if (intpGroup != null && intpGroup.getInterpreterHookRegistry() != null) {
-      registerHook(HookType.POST_EXEC_DEV, "z._displayhook()");
+      registerHook(HookType.POST_EXEC_DEV, "__zeppelin__._displayhook()");
     }
     DepInterpreter depInterpreter = getDepInterpreter();
 
@@ -130,7 +164,7 @@ public class PySparkInterpreter extends Interpreter implements ExecuteResultHand
             try {
               urlList.add(f.toURI().toURL());
             } catch (MalformedURLException e) {
-              logger.error("Error", e);
+              LOGGER.error("Error", e);
             }
           }
         }
@@ -147,7 +181,7 @@ public class PySparkInterpreter extends Interpreter implements ExecuteResultHand
             try {
               urlList.add(f.toURI().toURL());
             } catch (MalformedURLException e) {
-              logger.error("Error", e);
+              LOGGER.error("Error", e);
             }
           }
         }
@@ -161,20 +195,25 @@ public class PySparkInterpreter extends Interpreter implements ExecuteResultHand
       Thread.currentThread().setContextClassLoader(newCl);
       createGatewayServerAndStartScript();
     } catch (Exception e) {
-      logger.error("Error", e);
+      LOGGER.error("Error", e);
       throw new InterpreterException(e);
     } finally {
       Thread.currentThread().setContextClassLoader(oldCl);
     }
   }
 
-  private Map setupPySparkEnv() throws IOException{
+  private Map setupPySparkEnv() throws IOException {
     Map env = EnvironmentUtils.getProcEnvironment();
 
-    if (!env.containsKey("PYTHONPATH")) {
-      SparkConf conf = getSparkConf();
-      env.put("PYTHONPATH", conf.get("spark.submit.pyFiles").replaceAll(",", ":") +
-              ":../interpreter/lib/python");
+    // only set PYTHONPATH in local or yarn-client mode.
+    // yarn-cluster will setup PYTHONPATH automatically.
+    SparkConf conf = getSparkConf();
+    if (!conf.get("spark.submit.deployMode", "client").equals("cluster")) {
+      if (!env.containsKey("PYTHONPATH")) {
+        env.put("PYTHONPATH", PythonUtils.sparkPythonPath());
+      } else {
+        env.put("PYTHONPATH", PythonUtils.sparkPythonPath());
+      }
     }
 
     // get additional class paths when using SPARK_SUBMIT and not using YARN-CLIENT
@@ -189,7 +228,22 @@ public class PySparkInterpreter extends Interpreter implements ExecuteResultHand
       }
     }
 
+    LOGGER.info("PYTHONPATH: " + env.get("PYTHONPATH"));
     return env;
+  }
+
+  // Run python shell
+  // Choose python in the order of
+  // PYSPARK_DRIVER_PYTHON > PYSPARK_PYTHON > zeppelin.pyspark.python
+  public static String getPythonExec(Properties properties) {
+    String pythonExec = properties.getProperty("zeppelin.pyspark.python", "python");
+    if (System.getenv("PYSPARK_PYTHON") != null) {
+      pythonExec = System.getenv("PYSPARK_PYTHON");
+    }
+    if (System.getenv("PYSPARK_DRIVER_PYTHON") != null) {
+      pythonExec = System.getenv("PYSPARK_DRIVER_PYTHON");
+    }
+    return pythonExec;
   }
 
   private void createGatewayServerAndStartScript() {
@@ -201,22 +255,14 @@ public class PySparkInterpreter extends Interpreter implements ExecuteResultHand
     gatewayServer = new GatewayServer(this, port);
     gatewayServer.start();
 
-    // Run python shell
-    // Choose python in the order of
-    // PYSPARK_DRIVER_PYTHON > PYSPARK_PYTHON > zeppelin.pyspark.python
-    String pythonExec = getProperty("zeppelin.pyspark.python");
-    if (System.getenv("PYSPARK_PYTHON") != null) {
-      pythonExec = System.getenv("PYSPARK_PYTHON");
-    }
-    if (System.getenv("PYSPARK_DRIVER_PYTHON") != null) {
-      pythonExec = System.getenv("PYSPARK_DRIVER_PYTHON");
-    }
+    String pythonExec = getPythonExec(property);
+    LOGGER.info("pythonExec: " + pythonExec);
     CommandLine cmd = CommandLine.parse(pythonExec);
     cmd.addArgument(scriptPath, false);
     cmd.addArgument(Integer.toString(port), false);
     cmd.addArgument(Integer.toString(getSparkInterpreter().getSparkVersion().toNumber()), false);
     executor = new DefaultExecutor();
-    outputStream = new InterpreterOutputStream(logger);
+    outputStream = new InterpreterOutputStream(LOGGER);
     PipedOutputStream ps = new PipedOutputStream();
     in = null;
     try {
@@ -262,6 +308,10 @@ public class PySparkInterpreter extends Interpreter implements ExecuteResultHand
 
   @Override
   public void close() {
+    if (iPySparkInterpreter != null) {
+      iPySparkInterpreter.close();
+      return;
+    }
     executor.getWatchdog().destroyProcess();
     new File(scriptPath).delete();
     gatewayServer.shutdown();
@@ -275,10 +325,13 @@ public class PySparkInterpreter extends Interpreter implements ExecuteResultHand
   public class PythonInterpretRequest {
     public String statements;
     public String jobGroup;
+    public String jobDescription;
 
-    public PythonInterpretRequest(String statements, String jobGroup) {
+    public PythonInterpretRequest(String statements, String jobGroup,
+        String jobDescription) {
       this.statements = statements;
       this.jobGroup = jobGroup;
+      this.jobDescription = jobDescription;
     }
 
     public String statements() {
@@ -287,6 +340,10 @@ public class PySparkInterpreter extends Interpreter implements ExecuteResultHand
 
     public String jobGroup() {
       return jobGroup;
+    }
+
+    public String jobDescription() {
+      return jobDescription;
     }
   }
 
@@ -312,6 +369,7 @@ public class PySparkInterpreter extends Interpreter implements ExecuteResultHand
 
   public void setStatementsFinished(String out, boolean error) {
     synchronized (statementFinishedNotifier) {
+      LOGGER.debug("Setting python statement output: " + out + ", error: " + error);
       statementOutput = out;
       statementError = error;
       statementFinishedNotifier.notify();
@@ -324,12 +382,14 @@ public class PySparkInterpreter extends Interpreter implements ExecuteResultHand
   public void onPythonScriptInitialized(long pid) {
     pythonPid = pid;
     synchronized (pythonScriptInitializeNotifier) {
+      LOGGER.debug("onPythonScriptInitialized is called");
       pythonScriptInitialized = true;
       pythonScriptInitializeNotifier.notifyAll();
     }
   }
 
   public void appendOutput(String message) throws IOException {
+    LOGGER.debug("Output from python process: " + message);
     outputStream.getInterpreterOutput().write(message);
   }
 
@@ -337,9 +397,13 @@ public class PySparkInterpreter extends Interpreter implements ExecuteResultHand
   public InterpreterResult interpret(String st, InterpreterContext context) {
     SparkInterpreter sparkInterpreter = getSparkInterpreter();
     sparkInterpreter.populateSparkWebUrl(context);
-    if (sparkInterpreter.getSparkVersion().isUnsupportedVersion()) {
+    if (sparkInterpreter.isUnsupportedSparkVersion()) {
       return new InterpreterResult(Code.ERROR, "Spark "
           + sparkInterpreter.getSparkVersion().toString() + " is not supported");
+    }
+
+    if (iPySparkInterpreter != null) {
+      return iPySparkInterpreter.interpret(st, context);
     }
 
     if (!pythonscriptRunning) {
@@ -357,6 +421,7 @@ public class PySparkInterpreter extends Interpreter implements ExecuteResultHand
         try {
           pythonScriptInitializeNotifier.wait(1000);
         } catch (InterruptedException e) {
+          e.printStackTrace();
         }
       }
     }
@@ -390,10 +455,11 @@ public class PySparkInterpreter extends Interpreter implements ExecuteResultHand
       return new InterpreterResult(Code.ERROR, errorMessage);
     }
     String jobGroup = Utils.buildJobGroupId(context);
-    ZeppelinContext z = sparkInterpreter.getZeppelinContext();
-    z.setInterpreterContext(context);
-    z.setGui(context.getGui());
-    pythonInterpretRequest = new PythonInterpretRequest(st, jobGroup);
+    String jobDesc = "Started by: " + Utils.getUserName(context.getAuthenticationInfo());
+    SparkZeppelinContext __zeppelin__ = sparkInterpreter.getZeppelinContext();
+    __zeppelin__.setInterpreterContext(context);
+    __zeppelin__.setGui(context.getGui());
+    pythonInterpretRequest = new PythonInterpretRequest(st, jobGroup, jobDesc);
     statementOutput = null;
 
     synchronized (statementSetNotifier) {
@@ -425,22 +491,26 @@ public class PySparkInterpreter extends Interpreter implements ExecuteResultHand
 
   public void interrupt() throws IOException {
     if (pythonPid > -1) {
-      logger.info("Sending SIGINT signal to PID : " + pythonPid);
+      LOGGER.info("Sending SIGINT signal to PID : " + pythonPid);
       Runtime.getRuntime().exec("kill -SIGINT " + pythonPid);
     } else {
-      logger.warn("Non UNIX/Linux system, close the interpreter");
+      LOGGER.warn("Non UNIX/Linux system, close the interpreter");
       close();
     }
   }
 
   @Override
   public void cancel(InterpreterContext context) {
+    if (iPySparkInterpreter != null) {
+      iPySparkInterpreter.cancel(context);
+      return;
+    }
     SparkInterpreter sparkInterpreter = getSparkInterpreter();
     sparkInterpreter.cancel(context);
     try {
       interrupt();
     } catch (IOException e) {
-      logger.error("Error", e);
+      LOGGER.error("Error", e);
     }
   }
 
@@ -451,13 +521,20 @@ public class PySparkInterpreter extends Interpreter implements ExecuteResultHand
 
   @Override
   public int getProgress(InterpreterContext context) {
+    if (iPySparkInterpreter != null) {
+      return iPySparkInterpreter.getProgress(context);
+    }
     SparkInterpreter sparkInterpreter = getSparkInterpreter();
     return sparkInterpreter.getProgress(context);
   }
 
 
   @Override
-  public List<InterpreterCompletion> completion(String buf, int cursor) {
+  public List<InterpreterCompletion> completion(String buf, int cursor,
+      InterpreterContext interpreterContext) {
+    if (iPySparkInterpreter != null) {
+      return iPySparkInterpreter.completion(buf, cursor, interpreterContext);
+    }
     if (buf.length() < cursor) {
       cursor = buf.length();
     }
@@ -466,12 +543,11 @@ public class PySparkInterpreter extends Interpreter implements ExecuteResultHand
 
     //start code for completion
     SparkInterpreter sparkInterpreter = getSparkInterpreter();
-    if (sparkInterpreter.getSparkVersion().isUnsupportedVersion() == false
-            && pythonscriptRunning == false) {
+    if (sparkInterpreter.isUnsupportedSparkVersion() || pythonscriptRunning == false) {
       return new LinkedList<>();
     }
 
-    pythonInterpretRequest = new PythonInterpretRequest(completionCommand, "");
+    pythonInterpretRequest = new PythonInterpretRequest(completionCommand, "", "");
     statementOutput = null;
 
     synchronized (statementSetNotifier) {
@@ -485,13 +561,13 @@ public class PySparkInterpreter extends Interpreter implements ExecuteResultHand
         && pythonscriptRunning) {
         try {
           if (System.currentTimeMillis() - startTime > MAX_TIMEOUT_SEC * 1000) {
-            logger.error("pyspark completion didn't have response for {}sec.", MAX_TIMEOUT_SEC);
+            LOGGER.error("pyspark completion didn't have response for {}sec.", MAX_TIMEOUT_SEC);
             break;
           }
           statementFinishedNotifier.wait(1000);
         } catch (InterruptedException e) {
           // not working
-          logger.info("wait drop");
+          LOGGER.info("wait drop");
           return new LinkedList<>();
         }
       }
@@ -509,7 +585,7 @@ public class PySparkInterpreter extends Interpreter implements ExecuteResultHand
 
     List<InterpreterCompletion> results = new LinkedList<>();
     for (String name: completionList) {
-      results.add(new InterpreterCompletion(name, name));
+      results.add(new InterpreterCompletion(name, name, StringUtils.EMPTY));
     }
     return results;
   }
@@ -526,7 +602,7 @@ public class PySparkInterpreter extends Interpreter implements ExecuteResultHand
       completionScriptText = text.substring(0, cursor);
     }
     catch (Exception e) {
-      logger.error(e.toString());
+      LOGGER.error(e.toString());
       return null;
     }
     completionEndPosition = completionScriptText.length();
@@ -575,7 +651,22 @@ public class PySparkInterpreter extends Interpreter implements ExecuteResultHand
     return spark;
   }
 
-  public ZeppelinContext getZeppelinContext() {
+  private IPySparkInterpreter getIPySparkInterpreter() {
+    LazyOpenInterpreter lazy = null;
+    IPySparkInterpreter iPySpark = null;
+    Interpreter p = getInterpreterInTheSameSessionByClassName(IPySparkInterpreter.class.getName());
+
+    while (p instanceof WrappedInterpreter) {
+      if (p instanceof LazyOpenInterpreter) {
+        lazy = (LazyOpenInterpreter) p;
+      }
+      p = ((WrappedInterpreter) p).getInnerInterpreter();
+    }
+    iPySpark = (IPySparkInterpreter) p;
+    return iPySpark;
+  }
+
+  public SparkZeppelinContext getZeppelinContext() {
     SparkInterpreter sparkIntp = getSparkInterpreter();
     if (sparkIntp != null) {
       return getSparkInterpreter().getZeppelinContext();
@@ -636,12 +727,12 @@ public class PySparkInterpreter extends Interpreter implements ExecuteResultHand
   @Override
   public void onProcessComplete(int exitValue) {
     pythonscriptRunning = false;
-    logger.info("python process terminated. exit code " + exitValue);
+    LOGGER.info("python process terminated. exit code " + exitValue);
   }
 
   @Override
   public void onProcessFailed(ExecuteException e) {
     pythonscriptRunning = false;
-    logger.error("python process failed", e);
+    LOGGER.error("python process failed", e);
   }
 }
